Add LayoutContainer component and rename MainLayout; update routes to use LayoutContainer
