+++ packages/papaya-web/src/web/routes/_mainLayout.tsx
@@ -1,4 +1,4 @@
-import MainLayout from '@/layouts/main'
+import MainLayout from '@/layouts/MainLayout'
 import { createFileRoute, Outlet } from '@tanstack/react-router'
 
 export const Route = createFileRoute('/_mainLayout')({
