+++ packages/papaya-web/src/web/routes/_mainLayout/accounts.tsx
@@ -1,8 +1,13 @@
 import ManageAccounts from '@/components/journal/accounts/ManageAccounts'
+import LayoutContainer from '@/layouts/LayoutContainer'
 import { createFileRoute } from '@tanstack/react-router'
 
 const AccountsPage = () => {
-  return <ManageAccounts />
+  return (
+    <LayoutContainer>
+      <ManageAccounts />
+    </LayoutContainer>
+  )
 }
 
 export const Route = createFileRoute('/_mainLayout/accounts')({
