+++ packages/papaya-web/src/web/routes/_mainLayout/categories.tsx
@@ -1,8 +1,13 @@
 import ManageCategories from '@/components/journal/categories/ManageCategories'
+import LayoutContainer from '@/layouts/LayoutContainer'
 import { createFileRoute } from '@tanstack/react-router'
 
 const CategoriesPage = () => {
-  return <ManageCategories />
+  return (
+    <LayoutContainer>
+      <ManageCategories />
+    </LayoutContainer>
+  )
 }
 
 export const Route = createFileRoute('/_mainLayout/categories')({
